backend/requirements.txt
Flask==2.0.2
flask_marshmallow==0.14.0
Flask_SQLAlchemy==2.5.1
PyJWT==2.3.0
python-dotenv==0.19.2
Werkzeug==2.0.2
marshmallow-sqlalchemy==0.27.0